scripts/requirements.txt
torch>=2.0.0
onnx>=1.14.0
onnxruntime>=1.16.0
onnxruntime-extensions>=0.9.0
numpy>=1.24.0
Pillow>=10.0.0
